remove nameko and starlett/api files for python <3.6 (#763)
