+++ pytest.ini
@@ -1,37 +1,6 @@
 [pytest]
 addopts = --cov=scout_apm
           --cov-report=term-missing:skip-covered
-          # Disable Nameko's built-in Pytest plugin because it does
-          # eventlet.monkey_patch() early on, but we want to test things in a
-          # non-eventlet world!
-          # It lead to this failure:
-          # cls = <class 'eventlet.green.ssl.GreenSSLSocket'>, sock = <eventlet.greenio.base.GreenSocket object at 0x10667e588>, keyfile = None, certfile = None, server_side = False, cert_reqs = <VerifyMode.CERT_NONE: 0>
-          # ssl_version = <_SSLMethod.PROTOCOL_TLS: 2>, ca_certs = None, do_handshake_on_connect = True, args = ()
-          # kw = {'_context': <eventlet.green.ssl.GreenSSLContext object at 0x104d6f480>, 'server_hostname': 's3-us-west-1.amazonaws.com'}
-          #
-          #     def __new__(cls, sock=None, keyfile=None, certfile=None,
-          #                 server_side=False, cert_reqs=CERT_NONE,
-          #                 ssl_version=PROTOCOL_SSLv23, ca_certs=None,
-          #                 do_handshake_on_connect=True, *args, **kw):
-          #         if _is_under_py_3_7:
-          #             return super(GreenSSLSocket, cls).__new__(cls)
-          #         else:
-          #             if not isinstance(sock, GreenSocket):
-          #                 sock = GreenSocket(sock)
-          #             with _original_ssl_context():
-          #                 ret = _original_wrap_socket(
-          #                     sock=sock.fd,
-          #                     keyfile=keyfile,
-          #                     certfile=certfile,
-          #                     server_side=server_side,
-          #                     cert_reqs=cert_reqs,
-          #                     ssl_version=ssl_version,
-          #                     ca_certs=ca_certs,
-          #                     do_handshake_on_connect=False,
-          # >                   *args, **kw
-          #                 )
-          # E               TypeError: wrap_socket() got an unexpected keyword argument '_context'
-          -p no:pytest_nameko
 norecursedirs = src
 filterwarnings =
     error
