+++ tests/conftest.py
@@ -3,7 +3,6 @@ from __future__ import absolute_import, division, print_function, unicode_litera
 
 import logging
 import os
-import sys
 
 import celery
 import psutil
@@ -76,13 +75,6 @@ def caplog(caplog):
     caplog.set_level(logging.DEBUG)
     caplog.handler.addFilter(ExcludeDebugLogFilter("scout_test"))
     yield caplog
-
-
-# Some files are named to indicate they run on Python 3.6+ only (async
-# related) - ignore collecting them on older versions
-collect_ignore_glob = []
-if sys.version_info < (3, 6):
-    collect_ignore_glob.append("*_py36plus.py")
 
 
 class GlobalStateLeak(Exception):
