+++ tox.ini
@@ -32,10 +32,6 @@ deps =
     hug>=2.5.1
     httpretty
     jinja2
-    ; nameko tests temporarily disabled
-    ; Due to bad pinning of kombu only a past nameko version can be installed
-    ; alongside the latest celery.
-    ; nameko<3
     psutil
     pymongo
     pytest
